Add GenericEnqueuingOperation
