+++ hqscripts/generic_queue.py
@@ -1,0 +1,103 @@
+from datetime import datetime
+from time import sleep
+from optparse import make_option
+from django.core.management.base import BaseCommand, CommandError
+from django.conf import settings
+from dimagi.utils.parsing import string_to_datetime, json_format_datetime
+from dimagi.utils.couch.cache import cache_core
+from dimagi.utils.logging import notify_exception
+from redis_cache.cache import RedisCache
+
+class RedisClientError(Exception):
+    pass
+
+class GenericEnqueuingOperation(BaseCommand):
+    """
+    Implements a generic enqueuing operation.
+    """
+    args = ""
+    help = ""
+
+    def handle(self, *args, **options):
+        if self.use_queue():
+            self.validate_args(**options)
+            self.keep_fetching_items()
+
+    def keep_fetching_items(self):
+        while True:
+            try:
+                self.populate_queue()
+            except RedisClientError:
+                notify_exception(None,
+                    message="Could not get redis connection. Is redis up?")
+            except:
+                notify_exception(None,
+                    message="Could not populate %s." % self.get_queue_name())
+            sleep(15)
+
+    def populate_queue(self):
+        client = self.get_redis_client()
+        utcnow = datetime.utcnow()
+        entries = self.get_items_to_be_processed(utcnow)
+        for entry in entries:
+            queue_name = self.get_queue_name()
+            entry_id = entry["id"]
+            process_datetime = entry["key"]
+            enqueuing_lock = self.get_enqueuing_lock(client,
+                "%s-enqueuing-%s-%s" % (queue_name, entry_id, process_datetime))
+            if enqueuing_lock.acquire(blocking=False):
+                try:
+                    self.enqueue(entry_id)
+                except:
+                    # We couldn't enqueue, so release the lock
+                    enqueuing_lock.release()
+
+    def get_redis_client(self):
+        rcache = cache_core.get_redis_default_cache()
+        if not isinstance(rcache, RedisCache):
+            raise RedisClientError("Could not get redis connection.")
+        try:
+            client = rcache.raw_client
+        except:
+            raise RedisClientError("Could not get redis connection.")
+        return client
+
+    def get_enqueuing_lock(client, key):
+        lock_timeout = self.get_enqueuing_timeout()
+        return client.lock(key, timeout=lock_timeout)
+
+    def get_queue_name(self):
+        """Should return the name of this queue. Used for acquiring the
+        enqueuing lock to prevent enqueuing the same item twice"""
+        raise NotImplementedError("This method must be implemented.")
+
+    def get_enqueuing_timeout(self):
+        """Should return the timeout, in minutes, to use with the
+        enqueuing lock. This is essentially the number of minutes to
+        wait before enqueuing an unprocessed item again."""
+        raise NotImplementedError("This method must be implemented.")
+
+    def get_items_to_be_processed(self, utcnow):
+        """Should return the couch query result containing the items to be
+        enqueued. The result should just have the id of the item to be 
+        processed and the key from the couch view for each item. The couch 
+        view should emit a single value, which should be the timestamp that
+        the item should be processed. Since this just returns ids and keys, 
+        no limiting is necessary.
+        utcnow - The current timestamp, in utc, at the time of the method's
+            call. Retrieve all items to be processed before this timestamp."""
+        raise NotImplementedError("This method must be implemented.")
+
+    def enqueue(self, _id):
+        """This method should enqueue the item.
+        _id - The couch document _id of the item that is being referenced."""
+        raise NotImplementedError("This method must be implemented.")
+
+    def use_queue(self):
+        """If this is False, the handle() method will do nothing and return."""
+        return True
+
+    def validate_args(self, **options):
+        """Validate the options passed at the command line."""
+        pass
+
